Tracing layer: don't add virtual parent/child nodes into state
This situation can happen if a child/parent state requiring command is
performed on the active node while the active node is not on the current
section. If a parent or child is retrieved from the tracing layer state
and it is a virtual node its real parent or child will be used instead.

Thanks to @aschampion for the pointer.