Neuron navigator: update annotation filter node on annotation changes
While fixing catmaid/CATMAID#1330 it became apparent, that the
annotation filter node doesn't get update if annotations are added or
removed.

If an annotation filter is selected in the Neuron Navigator, it is now
updated on every annotation change event. This allows the UI to stay
consistent with model changes. Unfortunately, there is currently no easy
way to check if the annotation update would actually affect the node.
This, however, shouldn't be of big concern: the update is usually fast
and happens only if an annotation filter node is selected in the Neuron
Navigator, where you would expect such an update anyway.