Annotation search: don't use range parameters in request
Because of the way the Annotation Search widget is implemented, it can't
use partial results as they are returned by the back-end if range
parameters are used.

This problem was recently introduced by correcting a typo in the name of
one of the parameters. Before, parameter checking on the back-end
discarded the whole range request, because of the wrong parameter name.

Fixes catmaid/CATMAID#1375