Tracing layer: fix injection of active virtual nodes in node refresh
If the active node is virtual and sections are changed, its child and
parent are retrieved from the back-end along with all regular nodes for
the new section. Apart from its child and parent, the virtual node
itself has to be part of the new tracing layer state. Its execat
information like location is important, because it can't be
reconstructed from child and parent alone.